lambda/pyproject.toml
[tool.poetry]
name = "lambda"
version = "0.1.0"
description = ""
authors = ["luikymagno <luikymagno@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyJWT = "^2.8.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
